Split core extensions for plugin architecture
