+++ src/main/groovy/org/hidetake/groovy/ssh/extension/CoreExtensions.groovy
@@ -1,4 +1,4 @@
 package org.hidetake.groovy.ssh.extension
 
-trait CoreExtensions implements SudoExecution, SftpGet, SftpPut {
+trait CoreExtensions implements Command, BackgroundCommand, Shell, SudoExecution, SftpGet, SftpPut {
 }
