+++ src/main/groovy/org/hidetake/groovy/ssh/session/SessionHandler.groovy
@@ -44,135 +44,6 @@ class SessionHandler implements SessionExtension {
         operations.remote
     }
 
-    /**
-     * Performs a shell operation.
-     * This method blocks until channel is closed.
-     *
-     * @param settings shell settings
-     * @return output value of the command
-     */
-    void shell(HashMap settings) {
-        assert settings != null, 'settings must not be null'
-        log.info("Execute a shell with settings ($settings)")
-        operations.shell(operationSettings + new OperationSettings(settings))
-    }
-
-    /**
-     * Performs an execution operation.
-     * This method blocks until channel is closed.
-     *
-     * @param command
-     * @return output value of the command
-     */
-    String execute(String command) {
-        assert command, 'command must be given'
-        log.info("Execute a command ($command)")
-        operations.execute(operationSettings, command, null)
-    }
-
-    /**
-     * Performs an execution operation.
-     * This method blocks until channel is closed.
-     *
-     * @param command
-     * @param callback closure called with an output value of the command
-     * @return output value of the command
-     */
-    void execute(String command, Closure callback) {
-        assert command, 'command must be given'
-        assert callback, 'callback must be given'
-        log.info("Execute a command ($command) with callback")
-        operations.execute(operationSettings, command, callback)
-    }
-
-    /**
-     * Performs an execution operation.
-     * This method blocks until channel is closed.
-     *
-     * @param settings execution settings
-     * @param command
-     * @return output value of the command
-     */
-    String execute(HashMap settings, String command) {
-        assert command, 'command must be given'
-        assert settings != null, 'settings must not be null'
-        log.info("Execute a command ($command) with settings ($settings)")
-        operations.execute(operationSettings + new OperationSettings(settings), command, null)
-    }
-
-    /**
-     * Performs an execution operation.
-     * This method blocks until channel is closed.
-     *
-     * @param settings execution settings
-     * @param command
-     * @param callback closure called with an output value of the command
-     * @return output value of the command
-     */
-    void execute(HashMap settings, String command, Closure callback) {
-        assert command, 'command must be given'
-        assert callback, 'callback must be given'
-        assert settings != null, 'settings must not be null'
-        log.info("Execute a command ($command) with settings ($settings) and callback")
-        operations.execute(operationSettings + new OperationSettings(settings), command, callback)
-    }
-
-    /**
-     * Performs an execution operation.
-     * This method returns immediately and executes the command concurrently.
-     *
-     * @param command
-     */
-    void executeBackground(String command) {
-        assert command, 'command must be given'
-        log.info("Execute a command ($command) in background")
-        operations.executeBackground(operationSettings, command, null)
-    }
-
-    /**
-     * Performs an execution operation.
-     * This method returns immediately and executes the command concurrently.
-     *
-     * @param command
-     * @param callback closure called with an output value of the command
-     */
-    void executeBackground(String command, Closure callback) {
-        assert command, 'command must be given'
-        assert callback, 'callback must be given'
-        log.info("Execute a command ($command) with callback in background")
-        operations.executeBackground(operationSettings, command, callback)
-    }
-
-    /**
-     * Performs an execution operation.
-     * This method returns immediately and executes the command concurrently.
-     *
-     * @param settings execution settings
-     * @param command
-     */
-    void executeBackground(HashMap settings, String command) {
-        assert command, 'command must be given'
-        assert settings != null, 'settings must not be null'
-        log.info("Execute a command ($command) with settings ($settings) in background")
-        operations.executeBackground(operationSettings + new OperationSettings(settings), command, null)
-    }
-
-    /**
-     * Performs an execution operation.
-     * This method returns immediately and executes the command concurrently.
-     *
-     * @param settings execution settings
-     * @param command
-     * @param callback closure called with an output value of the command
-     */
-    void executeBackground(HashMap settings, String command, Closure callback) {
-        assert command, 'command must be given'
-        assert callback, 'callback must be given'
-        assert settings != null, 'settings must not be null'
-        log.info("Execute a command ($command) with settings ($settings) and callback in background")
-        operations.executeBackground(operationSettings + new OperationSettings(settings), command, callback)
-    }
-
     @Override
     void sftp(@DelegatesTo(SftpOperations) Closure closure) {
         assert closure, 'closure must be given'
